# app/backend/src/api-documentation/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Traveling Merchant",
    "version": "1",
    "description": "A back end application to keep track of transactions of a merchant buying product and selling product and keeping track of their inventory. All succesful operations start with \"Merchant: {merchantId}\" where merchantId is the id of the merchant that is logged in.",
    "license": {
      "name": "MIT",
      "url": "https://spdx.org/licenses/MIT.html"
    }
  },
  "servers": [
    {
      "url": "http://localhost:{PORT}",
      "variables": {},
      "description": "The application's URL is randomized after every initialization through its port number represented in the URL as {PORT}."
    }
  ],
  "paths": {
...
    "/transactions/{transactionId}": {
      "summary": "Request for a specific transaction's data",
      "description": "This path used to perform CRUD operations on transactions",
      "parameters": [
        {
          "in": "path",
          "name": "transactionId",
          "required": true,
          "schema": {
            "type": "integer",
            "minimum": 1
          },
          "description": "The transaction ID"
        }
      ],
      "get": {
        "summary": "Get transaction",
        "description": "This path uses the GET operation to get specific transactions.",
        "operationId": "",
        "responses": {
          "200": {
            "description": "OK, returns transaction data with associated ticket, and order.",
            "content": {
              "application/json": {
                "examples": {
                  "Transaction found given transaction id": {
                    "value": "{\"id\":4,\"ticketId\":2,\"orderId\":null,\"payment\":200,\"paymentType\":\"visa\",\"paidAt\":\"2025-01-09T00:00:00.000Z\",\"createdAt\":\"2025-01-09T00:00:00.000Z\",\"updatedAt\":\"2025-01-09T00:00:00.000Z\",\"ticket\":{\"id\":2,\"clientId\":3,\"cost\":488.52,\"paymentPlan\":\"biweekly\",\"description\":\"\",\"createdAt\":\"2025-01-09T00:00:00.000Z\",\"updatedAt\":\"2025-01-09T00:00:00.000Z\",\"paid\":200,\"returned\":0,\"owed\":288.52},\"order\":null}"
                  }
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND",
            "content": {
              "application/json": {
                "examples": {
                  "Transaction not found given transaction id": {
                    "value": "Transaction not found."
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Transactions"
        ]
      },
      "put": {
        "summary": "Update transaction",
        "description": "This path uses PUT operation to update transactions information.",
        "requestBody": {
          "description": "Transaction new information. The date properties (has the property name end with \"At\") should be in the following formats 2024-12, 2024-12-02, or \"2024-12-02T04:32:42.456Z\".",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/transactionsUpdates"
              }
            }
          }
        },
        "operationId": "",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "examples": {
                  "Updates one piece of information": {
                    "value": "Merchant: 1 transaction with id = 1 was updated with given payment = 647."
                  },
                  "Updates multiple pieces of transaction information": {
                    "value": "Merchant: 1 transaction with id = 4 was updated with given ticketId = 2, orderId = 7, and payment = 633."
                  },
                  "Updates all of an exisitng transaction information": {
                    "value": "Merchant: 1 transaction with id = 2 was updated with given ticketId = 5, orderId = 2, payment = 694, paymentType = advenio, and paidAt = Sat Dec 27 2025 09:33:10 GMT-0600 (Central Standard Time)."
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST",
            "content": {
              "application/json": {
                "examples": {
                  "No inputs are given in the request body": {
                    "value": "Bad input request."
                  }
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND",
            "content": {
              "application/json": {
                "examples": {
                  "Transaction not found given transaction id": {
                    "value": "Transaction not found."
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Transactions"
        ]
      },
      "delete": {
        "summary": "Delete transaction",
        "description": "This path uses the DELETE operation to delete transactions from database.",
        "operationId": "",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "examples": {
                  "Transaction is succesfully deleted": {
                    "value": "Merchat: 1 has deleted a transaction with id = 3."
                  }
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND",
            "content": {
              "application/json": {
                "examples": {
                  "Transaction not found given transaction id": {
                    "value": "Transaction not found."
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Transactions"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "transactionsUpdates": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1
          },
          "ticketId": {
            "type": "integer",
            "minimum": 1
          },
          "orderId": {
            "type": "integer",
            "minimum": 1
          },
          "payment": {
            "type": "number",
            "minimum": 0
          },
          "paymentType": {
            "type": "string"
          },
          "paidAt": {
            "type": "string"
          }
        },
        "xml": {
          "name": "Updated transaction information"
        }
      },
...
    }
  }
}